chore: add utility getNodesWithDocClientNamedImportFromDeepPath
